requirements.txt
python-telegram-bot==10.1.0
peewee==3.6.4
python-decouple==3.1